created produc reader module
