+++ dnora/read/product_readers.py
@@ -1,0 +1,146 @@
+from .abstract_readers import SpectralDataReader
+from dnora.read.file_structure import FileStructure
+from dnora.read.product_configuration import ProductConfiguration
+from functools import partial
+from dnora.type_manager.data_sources import DataSource
+from dnora.read.ds_read_functions import (
+    read_ds_list,
+    read_first_ds,
+    find_time_var_in_ds,
+)
+from geo_skeletons import PointSkeleton
+from dnora import msg
+import xarray as xr
+
+
+class SpectralProductReader(SpectralDataReader):
+    """This should not be used directly, but is a template for model specific implementations"""
+
+    # This defines filenames, data sources etc.
+    product_configuration = ProductConfiguration(default_data_source=DataSource.LOCAL)
+
+    # This defines how the file structure of the model output is set up
+    file_structure = FileStructure(
+        stride="month",
+        hours_per_file=None,
+        last_file="",
+        lead_time=0,
+        offset=0,
+    )
+
+    @staticmethod
+    def returning_ds() -> bool:
+        return True
+
+    def __init__(
+        self,
+        stride: (
+            int | str | None
+        ) = None,  # Integer is number of hours, 'month' for monthly files
+        hours_per_file: int | None = None,  # None for stride = 'month'
+        last_file: str | None = None,
+        lead_time: int | None = None,
+        offset: int | None = None,
+        tile: str | None = None,
+    ) -> None:
+        if stride is not None:
+            self.file_structure.stride = stride
+
+        if hours_per_file is not None:
+            self.file_structure.hours_per_file = hours_per_file
+
+        if last_file is not None:
+            self.file_structure.last_file = last_file
+
+        if lead_time is not None:
+            self.file_structure.lead_time = lead_time
+
+        if offset is not None:
+            self.file_structure.offset = offset
+
+        self.set_convention(self.product_configuration.convention)
+
+        self._default_filename = self.product_configuration.filename
+        self._default_filenames = self.product_configuration.default_filenames
+        self._default_folders = self.product_configuration.default_folders
+        self.set_default_data_source(self.product_configuration.default_data_source)
+
+        self._tile = tile or self.product_configuration.tile
+
+    def get_coordinates(
+        self,
+        grid,
+        start_time,
+        source: DataSource,
+        folder: str,
+        filename: str,
+        tile: str | None = None,
+        **kwargs,
+    ) -> dict:
+        """Reads first time instance of first file to get longitudes and latitudes for the PointPicker"""
+        tile = tile or self._tile
+        tile_name = self.product_configuration.tile_names.get(tile)
+
+        folder = self._folder(folder, source, tile=tile, tile_name=tile_name)
+        filename = self._filename(filename, source, tile=tile, tile_name=tile_name)
+        ds = read_first_ds(folder, filename, start_time, self.file_structure)
+        points = PointSkeleton.from_ds(ds)
+        return points.coord_dict()
+
+    def __call__(
+        self,
+        obj_type,
+        grid,
+        start_time,
+        end_time,
+        source: DataSource,
+        folder: str,
+        filename: str,
+        inds,
+        dnora_class=None,
+        tile: str | None = None,
+        **kwargs,
+    ) -> tuple[dict]:
+        """Reads in all boundary spectra between the given times and at for the given indeces"""
+        tile = tile or self._tile
+        tile_name = self.product_configuration.tile_names.get(tile)
+        folder = self._folder(folder, source, tile=tile, tile_name=tile_name)
+        filename = self._filename(filename, source, tile=tile, tile_name=tile_name)
+        start_times, end_times, file_times = self.file_structure.create_time_stamps(
+            start_time, end_time
+        )
+        msg.info(
+            f"Getting boundary spectra from {self.name()} from {start_time} to {end_time}"
+        )
+
+        ds_creator_function = partial(
+            self.product_configuration.ds_creator_function, inds=inds
+        )
+        ds_list = read_ds_list(
+            start_times,
+            end_times,
+            file_times,
+            folder,
+            filename,
+            ds_creator_function,
+        )
+
+        msg.info("Merging dataset together (this might take a while)...")
+        time_var = self.product_configuration.time_var or find_time_var_in_ds(
+            ds_list[0]
+        )
+        ds = xr.concat(ds_list, dim=time_var)
+
+        ds_aliases = self.product_configuration.ds_aliases
+        core_aliases = self.product_configuration.get_core_aliases(obj_type)
+
+        points = dnora_class.from_ds(
+            ds,
+            dynamic=False,
+            ignore_vars=[],
+            only_vars=[],
+            ds_aliases=ds_aliases,
+            core_aliases=core_aliases,
+        )
+
+        return points.ds()
