+++ dnora/read/spectra/__init__.py
@@ -1,2 +1,2 @@
-from .spectral_readers import SWAN_Ascii, SpectralProductReader
+from .spectral_readers import SWAN_Ascii
 from . import metno, ec  # , nchmf
